+++ frontend/util/status.js
@@ -1,3 +1,3 @@
 export function is_leader(status) {
-    return status && status.election && status.election.is_leader;
+    return status && status.election && status.election.isLeader;
 }